# stx/dotnet/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Styx Job Manager API",
    "version": "1.0.0",
    "description": "REST API for the Styx actuarial job management platform.\n\nThree controllers:\n- **JobManager** — job initialisation data and submission\n- **JobGrid**    — job monitoring, children hierarchy, task logs, cancel/requeue\n- **JobEnvironment** — UNC filesystem exploration\n- **TaskFlow**   — task state notifications\n"
  },
  "servers": [
    {
      "url": "{baseUrl}",
      "description": "Configurable base URL (set in Settings)",
      "variables": {
        "baseUrl": {
          "default": "http://localhost:5000"
        }
      }
    }
  ],
...
  "paths": {
...
    "/api/JobManager/refreshinputs": {
      "get": {
        "tags": [
          "JobManager"
        ],
        "summary": "Refresh input folders for an environment path",
        "parameters": [
          {
            "name": "environment",
            "in": "query",
            "required": false,
            "description": "UNC path to the environment root",
            "schema": {
              "type": "string",
              "example": "\\\\srvrbxassufp01\\hpc\\Cas01\\4-Omen"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "inputs": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/JobManager/submit/savings": {
      "post": {
        "tags": [
          "JobManager"
        ],
        "summary": "Submit a Savings (Omen Épargne) job",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavingsSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/JobManager/submit/nonlife": {
      "post": {
        "tags": [
          "JobManager"
        ],
        "summary": "Submit a NonLife job",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NonLifeSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/JobManager/submit/risklife": {
      "post": {
        "tags": [
          "JobManager"
        ],
        "summary": "Submit a RiskLife job",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskLifeSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/JobManager/submit/risklifekp": {
      "post": {
        "tags": [
          "JobManager"
        ],
        "summary": "Submit a RiskLifeKP job",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskLifeSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/JobManager/submit/tdr": {
      "post": {
        "tags": [
          "JobManager"
        ],
        "summary": "Submit a TdR job",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TdrSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/JobManager/submit/brd": {
      "post": {
        "tags": [
          "JobManager"
        ],
        "summary": "Submit a BRD job",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrdSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/JobManager/submit/ufx": {
      "post": {
        "tags": [
          "JobManager"
        ],
        "summary": "Submit a UFX (Universal Format Exchange) job",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UfxSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/JobManager/submit/custominput": {
      "post": {
        "tags": [
          "JobManager"
        ],
        "summary": "Submit a Custom Input job",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomInputSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/JobManager/submit/scenarioTransfo": {
      "post": {
        "tags": [
          "JobManager"
        ],
        "summary": "Submit a Scenario Transformator job",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioTransfoSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/JobGrid/jobs/children": {
      "get": {
        "tags": [
          "JobGrid"
        ],
        "summary": "Get child job groups for a parent job",
        "parameters": [
          {
            "name": "jobId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "example": 1134535
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobChildrenResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/JobGrid/jobs/{jobId}/tasks": {
      "get": {
        "tags": [
          "JobGrid"
        ],
        "summary": "Get task-level log for a leaf job",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "example": 1134538
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobTasksResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/JobGrid/cancel": {
      "post": {
        "tags": [
          "JobGrid"
        ],
        "summary": "Cancel a job",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "jobId"
                ],
                "properties": {
                  "jobId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusMessage"
                }
              }
            }
          }
        }
      }
    },
    "/api/JobGrid/requeue": {
      "post": {
        "tags": [
          "JobGrid"
        ],
        "summary": "Requeue a failed or cancelled job",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "jobId"
                ],
                "properties": {
                  "jobId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusMessage"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "JobState": {
        "type": "string",
        "enum": [
          "Running",
          "Queued",
          "Finished",
          "Failed",
          "Canceled"
        ],
        "description": "State of a parent job (string form used in JobGrid.jobs)"
      },
      "ChildJobStateCode": {
        "type": "integer",
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "description": "Numeric state used in child jobs / job groups.\n0 = Queued, 1 = Running, 2 = Finished, 3 = Failed, 4 = Canceled\n"
      },
      "JobPriority": {
        "type": "string",
        "enum": [
          "Normal",
          "High",
          "BelowNormal",
          "AboveNormal",
          "Automatic"
        ]
      },
      "StatusMessage": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "OK"
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "TaskIdEntry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "gridId": {
            "type": "integer"
          }
        }
      },
      "ExternalIdEntry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "bpcId": {
            "type": "integer"
          },
          "xGridId": {
            "type": "integer"
          }
        }
      },
      "LightGridResult": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "integer"
          },
          "tasksIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskIdEntry"
            }
          }
        }
      },
      "ExternalApiResult": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExternalIdEntry"
            }
          },
          "status": {
            "type": "string",
            "example": "Submitted"
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ApiResults": {
        "type": "object",
        "properties": {
          "lightGridSubmissionResult": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LightGridResult"
            }
          },
          "externalApiSubmissionResult": {
            "$ref": "#/components/schemas/ExternalApiResult"
          }
        }
      },
      "SubmitResponse": {
        "type": "object",
        "properties": {
          "parentId": {
            "type": "integer",
            "example": 1001
          },
          "launchedJobCount": {
            "type": "integer",
            "example": 1
          },
          "apiResults": {
            "$ref": "#/components/schemas/ApiResults"
          }
        }
      },
...
      "OmenSubmitBase": {
        "description": "Fields common to all Omen (savings / risklife / nonlife / tdr / brd) submissions.",
        "type": "object",
        "required": [
          "environment",
          "inputs",
          "refModelSelected",
          "model",
          "version"
        ],
        "properties": {
          "jobName": {
            "type": "string"
          },
          "environment": {
            "type": "string",
            "description": "UNC path to the environment root",
            "example": "\\\\srvrbxassufp01\\hpc\\Cas01\\4-Omen"
          },
          "inputs": {
            "type": "string",
            "description": "Input folder name relative to environment/input",
            "example": "Input23.01.00_2022Y_vfinale"
          },
          "refModelSelected": {
            "type": "boolean",
            "description": "true = reference model; false = custom version path"
          },
          "model": {
            "type": "string",
            "example": "2026.01.00"
          },
          "version": {
            "type": "string",
            "example": "2026.01.00_prod"
          },
          "advOptions": {
            "type": "boolean"
          }
        }
      },
      "SavingsSubmitRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/OmenSubmitBase"
          },
          {
            "type": "object",
            "properties": {
              "detEnabled": {
                "type": "boolean"
              },
              "detRange": {
                "type": "string",
                "example": "1-2000"
              },
              "detPeriod": {
                "type": "string",
                "example": "1y"
              },
              "detMonths": {
                "type": "integer",
                "example": 12
              },
              "stoEnabled": {
                "type": "boolean"
              },
              "stoRange": {
                "type": "string",
                "example": "1-3000"
              },
              "stoPeriod": {
                "type": "string",
                "example": "5y"
              },
              "stoMonths": {
                "type": "integer",
                "example": 60
              },
              "pricerEnabled": {
                "type": "boolean"
              },
              "pricerRange": {
                "type": "string",
                "example": "1-4000"
              },
              "pricerPeriod": {
                "type": "string",
                "example": "30y"
              },
              "pricerMonths": {
                "type": "integer",
                "example": 360
              },
              "guaranteedFloor": {
                "type": "boolean"
              },
              "omenType": {
                "type": "string",
                "enum": [
                  "S2",
                  "S1",
                  "FR_GAAP",
                  "Solvency_II"
                ]
              },
              "scenarios": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "Selected scenario numbers (as strings)",
                "example": [
                  "1",
                  "2"
                ]
              }
            }
          }
        ]
      },
      "NonLifeSubmitRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/OmenSubmitBase"
          },
          {
            "type": "object",
            "properties": {
              "period": {
                "type": "integer",
                "description": "Projection period in months",
                "example": 80
              }
            }
          }
        ]
      },
      "RiskLifeSubmitRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/OmenSubmitBase"
          },
          {
            "type": "object",
            "properties": {
              "period": {
                "type": "integer",
                "description": "Projection period in months",
                "example": 360
              },
              "scenarios": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "Selected scenario numbers (as strings)",
                "example": [
                  "1",
                  "2",
                  "3"
                ]
              }
            }
          }
        ]
      },
      "TdrSubmitRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/OmenSubmitBase"
          },
          {
            "type": "object",
            "properties": {
              "period": {
                "type": "integer",
                "description": "Period in months",
                "example": 9
              }
            }
          }
        ]
      },
      "BrdSubmitRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/OmenSubmitBase"
          },
          {
            "type": "object",
            "properties": {
              "projectionDuration": {
                "type": "integer",
                "description": "Projection duration in years",
                "example": 30
              },
              "scenarios": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "1",
                  "2",
                  "3",
                  "4"
                ]
              }
            }
          }
        ]
      },
      "UfxSubmitRequest": {
        "type": "object",
        "required": [
          "path"
        ],
        "properties": {
          "jobName": {
            "type": "string"
          },
          "path": {
            "type": "string",
            "description": "UNC path to the UFX file or folder"
          },
          "isFolder": {
            "type": "boolean",
            "description": "true if path points to a folder (batch UFX)"
          }
        }
      },
      "CustomInputSubmitRequest": {
        "type": "object",
        "required": [
          "inputsFolder",
          "actionsFolder"
        ],
        "properties": {
          "inputsFolder": {
            "type": "string",
            "description": "UNC path to the inputs folder"
          },
          "actionsFolder": {
            "type": "string",
            "description": "UNC path to the custom actions folder"
          }
        }
      },
      "ScenarioTransfoSubmitRequest": {
        "type": "object",
        "required": [
          "environment",
          "modelType",
          "periode",
          "iterations",
          "coupons"
        ],
        "properties": {
          "environment": {
            "type": "string",
            "description": "UNC path to the data folder to transform"
          },
          "modelType": {
            "type": "string",
            "enum": [
              "2017",
              "2018",
              "2020",
              "2020v2"
            ]
          },
          "periode": {
            "type": "string",
            "enum": [
              "mois",
              "annee"
            ]
          },
          "iterations": {
            "type": "integer",
            "example": 1000
          },
          "coupons": {
            "type": "string",
            "enum": [
              "ZC",
              "TC"
            ]
          },
          "split": {
            "type": "boolean",
            "description": "Split output into per-scenario files"
          }
        }
      },
...
      "ChildJob": {
        "description": "Leaf job inside a JobGroup.",
        "type": "object",
        "required": [
          "id",
          "state",
          "name",
          "progress"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "state": {
            "$ref": "#/components/schemas/ChildJobStateCode"
          },
          "name": {
            "type": "string"
          },
          "progress": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "priority": {
            "$ref": "#/components/schemas/JobPriority"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "submitted": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time"
          },
          "environment": {
            "type": "string"
          }
        }
      },
      "JobGroup": {
        "description": "A sequential execution phase within a parent job.",
        "type": "object",
        "required": [
          "id",
          "status",
          "name",
          "progress"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/ChildJobStateCode"
          },
          "name": {
            "type": "string",
            "example": "Run sto_alm_16"
          },
          "progress": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "updated": {
            "type": "string",
            "format": "date-time"
          },
          "parentId": {
            "type": "integer"
          },
          "launchedJobCount": {
            "type": "integer"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChildJob"
            }
          }
        }
      },
      "JobChildrenResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "OK"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "warnings": {
            "type": "string",
            "nullable": true
          },
          "jobId": {
            "type": "integer"
          },
          "jobGroups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobGroup"
            }
          }
        }
      },
      "Task": {
        "description": "Individual command within a leaf job.",
        "type": "object",
        "required": [
          "id",
          "state",
          "command"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "command": {
            "type": "string",
            "example": "StyxLauncher.Cmd.exe -runScenario sto_alm_16_batch_1 1-500"
          },
          "output": {
            "type": "string",
            "nullable": true
          },
          "startTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "endTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "JobTasksResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "OK"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "warnings": {
            "type": "string",
            "nullable": true
          },
          "jobId": {
            "type": "integer"
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Task"
            }
          }
        }
      },
...
    }
  }
}